Fix file modes for Python 3
Encrypted data files need to be in binary mode, and non-encrypted CSV
files need to be in text mode under Python 3.

This change may break support for Windows as the controller, but we
don't currently have any way to test that, so it's probably broken
anyway and definitely unsupported. If we do get a way to test Windows,
this change doesn't make it any harder to use the OS to choose the
file mode.