Better error messages for missing programs
Check whether programs we need (like dmidecode and tune2fs) are
present on the remote machine before running commands that depend on
them. If they aren't present, we give report messages like "N/A
(dmidecode not found)" instead of propagating a shell error message to
the output.

This also adds a file to the documentation listing exactly what
executables are needed for each set of facts. We do not explicitly
check for standard Unix things like grep and sed, so if those are
missing we would still give a shell error, but the documentation lists
all of the programs we need.