Merge 0.30 version into master branch (#451)
* Factor out Ansible utils and host discovery

Extract two separate modules from rho/scancommand.py. This is part of
the preparation for scanning for Fuse.

* Move inventory scan to its own module

Organize the scanning code and make it more clear what's happening.

* Minor cleanups to inventory scanning

Part of a larger set of changes around reorganizing that code.

* Improve error messages in ansible_utils.py

* Fix lint errors

* Use 'locate' command to find jboss-modules.jar

This lets us scan the whole filesystem (or at least the locate
database) when it's been preindexed.

* Review fixes

Add locate to the 'jboss' fact group and update the fact docs. Thanks
to chambrid for the review.

* Look for init files related to JBoss EAP

Add another indicator to the JBoss EAP scanning. Now, Rho will look
for init files and system services related to JBoss in addition to
everything else.

* Prep dev branch with new version before merging dev into master. (#450)

* Add changelog updates. Closes #438. (#439)

* Fix spelling error in issue template